requirements/base.txt
dj-database-url==2.0.0
Django==4.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.6
python-decouple==3.8
pytz==2023.3
